add template of dummy
