+++ next-app/pages/dummy/index.tsx
@@ -1,15 +1,12 @@
 // import react or next
 import { useEffect } from 'react';
 import type { GetStaticProps, NextPage } from 'next';
-import { useRouter } from 'next/router';
 
-// import store
-import { useAppSelector } from '@/store';
-import { init, clear, update } from '@/usecases/dummy';
+import { init } from '@/usecases/dummy';
 
 // import components
-import { Layout } from '@/components/templates/layout';
-import { Template } from '@/components/templates/template';
+import { NavbarFooterLayoutTemplate } from '@/components/templates/navbar-footer-layout';
+import { DummyTemplate } from '@/components/templates/dummy';
 
 export const getStaticProps: GetStaticProps = async () => {
   const developerNames = ['asamin', 'sugya'];
@@ -29,19 +26,6 @@ const DummyPage: NextPage<Props> = (props: Props) => {
   const title = 'Dummy Page - Open Camp';
   const description = 'dummy page of open camp';
 
-  const router = useRouter();
-  const { name } = useAppSelector((state) => state['ui/dummy']);
-
-  const handleUpdate = async () => {
-    await update(`DummyState updated by ${developerNames.join('&')}`);
-  };
-  const handleClearButton = async () => {
-    await clear();
-  };
-  const goBackPage = () => {
-    router.back();
-  };
-
   useEffect(() => {
     async function handleInit() {
       await init();
@@ -50,32 +34,9 @@ const DummyPage: NextPage<Props> = (props: Props) => {
   }, []);
 
   return (
-    <Layout title={title} description={description}>
-      <main>
-        <Template />
-        <h1>Welcome to Dummy Page!</h1>
-        <h3>{name}</h3>
-        <button
-          type="button"
-          onClick={async () => {
-            handleUpdate();
-          }}
-        >
-          update state name
-        </button>
-        <button
-          type="button"
-          onClick={async () => {
-            handleClearButton();
-          }}
-        >
-          clear state name
-        </button>
-        <button type="button" onClick={goBackPage}>
-          back home
-        </button>
-      </main>
-    </Layout>
+    <NavbarFooterLayoutTemplate title={title} description={description}>
+      <DummyTemplate developerNames={developerNames} />
+    </NavbarFooterLayoutTemplate>
   );
 };
